ai-retriever/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
faiss-cpu==1.8.0.post1
numpy==1.26.4
python-dotenv==1.0.1
pandas==2.2.2
pyarrow==16.1.0
tqdm==4.66.4
